removed superfluous static_cast for facing

+++ src/mower.cpp
@@ -67,7 +67,7 @@ int                 Mower::getSteps() const { return steps_; }
 
 int                 Mower::getTurns() const { return turns_; }
 
-std::string         Mower::getMowerString() const { return MOWER_STRINGS[static_cast<int>(facing_)]; }
+std::string         Mower::getMowerString() const { return MOWER_STRINGS[facing_]; }
 
 std::string         Mower::getFacingString() const { return Facing::toString(facing_); }
